requirements.txt
# Core ML and Data Science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
tensorflow>=2.10.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Image Processing
opencv-python>=4.6.0
Pillow>=9.0.0
rasterio>=1.3.0

# Web Frameworks
fastapi>=0.95.0
uvicorn>=0.20.0
streamlit>=1.20.0
streamlit-folium>=0.11.0

# Data APIs and Processing
requests>=2.28.0
earthengine-api>=0.1.350
google-auth>=2.16.0
google-cloud-storage>=2.7.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.64.0

# Geographic and Geospatial
geopandas>=0.12.0
folium>=0.14.0
pyproj>=3.4.0

# Testing and Development
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Jupyter and Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.5.0

# Time Series and Statistical Analysis
statsmodels>=0.13.0
scipy>=1.9.0

# Model Deployment and Monitoring
joblib>=1.2.0
mlflow>=2.0.0

# Additional utilities
schedule>=1.2.0
python-multipart>=0.0.5
jinja2>=3.1.0